Android and ios releated issues are resolved

+++ src/3d_shape/surface/base_surface/BaseSurface.h
@@ -21,7 +21,7 @@ public:
 
   unsigned long getEdgeByIndex(short index);
 
-  bool areEdgesDataValid(int nodesSize);
+  bool areEdgesDataValid(unsigned long nodesSize);
 
   virtual ~BaseSurface() = default;
